adding more fields to ticket and changing nft price field to float

+++ core/models.py
@@ -15,7 +15,7 @@ class NFT(models.Model):
     owner = models.ForeignKey(User, null=False, blank=False, on_delete=models.CASCADE)
     creator = models.CharField(max_length=100, null=False, blank=False)
     date = models.DateTimeField(verbose_name="تاریخ", auto_now=True)
-    last_price = models.IntegerField(verbose_name='آخرین قیمت', null=False, blank=False)
+    last_price = models.FloatField(verbose_name='آخرین قیمت', null=False, blank=False)
     image_url = models.TextField(default='data:image', null=False, blank=False)
     start_date = models.DateTimeField(verbose_name='تاریخ شروع مزایده', null=True, blank=True)
     end_date = models.DateTimeField(verbose_name='تاریخ پایان مزایده', null=True, blank=True)
